requirements.txt
pytest-asyncio==0.24.0
pytest-retry==1.6.3
pytest-mock==3.14.0